chore(PBF): refactor modules based on package-by-feature
