+++ app/src/main/java/com/andela/android/javadevelopers/service/DevelopersApi.java
@@ -1,6 +1,6 @@
 package com.andela.android.javadevelopers.service;
 
-import com.andela.android.javadevelopers.model.DevelopersListResponse;
+import com.andela.android.javadevelopers.home.model.DevelopersListResponse;
 
 import retrofit2.Call;
 import retrofit2.http.GET;
