+++ app/src/main/res/layout/activity_details.xml
@@ -6,7 +6,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.DetailsActivity">
+    tools:context=".details.view.DetailsActivity">
 
     <android.support.design.widget.AppBarLayout
         android:id="@+id/appbar"
