+++ app/src/main/res/layout/activity_location.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.LocationActivity"
+    tools:context=".location.view.LocationActivity"
     tools:layout_editor_absoluteY="81dp">
 
     <Spinner
