+++ app/src/main/res/layout/activity_main.xml
@@ -7,7 +7,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:padding="6dp"
-        tools:context=".view.MainActivity">
+        tools:context=".home.view.MainActivity">
 
         <android.support.v7.widget.RecyclerView
             android:id="@+id/recyclerView"
