+++ app/src/main/res/menu/share.xml
@@ -2,7 +2,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".view.DetailsActivity">
+    tools:context=".details.view.DetailsActivity">
     <item
         android:id="@+id/share_action_button"
         android:icon="@drawable/abc_ic_menu_share_mtrl_alpha"
